requirements.txt
django
qdrant-client
osxphotos
pillow